pyproject.toml
[project]
name = "ai-travell-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
]
